Move Accept-Encoding parser to utils and harden qvalue handling
Relocate parse_accept_encoding from responders to utils for reuse, and
add edge-case tests covering whitespace, case, multiple parameters and
invalid qvalues. Reject non-finite and out-of-range qvalues per RFC 9110.
