+++ src/servestatic/utils.py
@@ -4,6 +4,7 @@ import asyncio
 import concurrent.futures
 import contextlib
 import functools
+import math
 import os
 from concurrent.futures import ThreadPoolExecutor
 from typing import TYPE_CHECKING, Concatenate, ParamSpec, TypeVar, cast
@@ -36,6 +37,45 @@ def decode_path_info(path_info: str) -> str:
 def ensure_leading_trailing_slash(path: str | None) -> str:
     path = (path or "").strip("/")
     return f"/{path}/" if path else "/"
+
+
+def parse_accept_encoding(header: str) -> set[str]:
+    """
+    Return the set of content codings the client is willing to accept.
+
+    The header is parsed as described in RFC 9110 section 12.5.3 rather than
+    matched as a substring, so that a coding the client has refused with
+    ``q=0`` is not offered to it, and a coding name is only recognised when it
+    appears as a whole token.
+    """
+    accepted: set[str] = set()
+    for entry in header.split(","):
+        coding, _, params = entry.partition(";")
+        coding = coding.strip().lower()
+        if not coding:
+            continue
+        quality = 1.0
+        for param in params.split(";"):
+            name, sep, value = param.partition("=")
+            if name.strip().lower() != "q":
+                continue
+            if not sep:
+                break
+            try:
+                quality = float(value.strip())
+            except ValueError:
+                # An unparseable qvalue makes the whole entry invalid; treat
+                # the coding as unacceptable rather than guess.
+                quality = 0.0
+            else:
+                # Per RFC 9110 a qvalue must be a number in the range 0.000 to
+                # 1.000; any other finite value (e.g. ``inf``) is invalid.
+                if not math.isfinite(quality) or not 0.0 <= quality <= 1.0:
+                    quality = 0.0
+            break
+        if quality > 0:
+            accepted.add(coding)
+    return accepted
 
 
 def scantree(root: str | PathLike[str]) -> Iterator[tuple[str, os.stat_result]]:
